+++ packages/block/src/header/constructors.ts
@@ -2,7 +2,7 @@ import { RLP } from '@tvmjs/rlp'
 import { EthereumJSErrorWithoutCode, bigIntToBytes, equalsBytes } from '@tvmjs/util'
 
 import { generateCliqueBlockExtraData } from '../consensus/clique.ts'
-import { numberToHex, rejectBlobFields, valuesArrayToHeaderData } from '../helpers.ts'
+import { numberToHex, rejectRemovedHeaderFields, valuesArrayToHeaderData } from '../helpers.ts'
 import { BlockHeader } from '../index.ts'
 
 import type { BlockHeaderBytes, BlockOptions, HeaderData, JSONRPCBlock } from '../types.ts'
@@ -25,7 +25,7 @@ export function createBlockHeader(headerData: HeaderData = {}, opts: BlockOption
  */
 export function createBlockHeaderFromBytesArray(values: BlockHeaderBytes, opts: BlockOptions = {}) {
   const headerData = valuesArrayToHeaderData(values)
-  const { number, baseFeePerGas, parentBeaconBlockRoot, requestsHash } = headerData
+  const { number, baseFeePerGas, requestsHash } = headerData
   const header = createBlockHeader(headerData, opts)
   if (header.common.isActivatedEIP(1559) && baseFeePerGas === undefined) {
     const eip1559ActivationBlock = bigIntToBytes(header.common.eipBlock(1559)!)
@@ -36,10 +36,6 @@ export function createBlockHeaderFromBytesArray(values: BlockHeaderBytes, opts: 
       throw EthereumJSErrorWithoutCode('invalid header. baseFeePerGas should be provided')
     }
   }
-  if (header.common.isActivatedEIP(4788) && parentBeaconBlockRoot === undefined) {
-    throw EthereumJSErrorWithoutCode('invalid header. parentBeaconBlockRoot should be provided')
-  }
-
   if (header.common.isActivatedEIP(7685) && requestsHash === undefined) {
     throw EthereumJSErrorWithoutCode('invalid header. requestsHash should be provided')
   }
@@ -102,7 +98,7 @@ export function createSealedCliqueBlockHeader(
  * @param options - An object describing the blockchain
  */
 export function createBlockHeaderFromRPC(blockParams: JSONRPCBlock, options?: BlockOptions) {
-  rejectBlobFields(blockParams)
+  rejectRemovedHeaderFields(blockParams)
   const {
     parentHash,
     sha3Uncles,
@@ -121,7 +117,6 @@ export function createBlockHeaderFromRPC(blockParams: JSONRPCBlock, options?: Bl
     nonce,
     baseFeePerGas,
     withdrawalsRoot,
-    parentBeaconBlockRoot,
     requestsHash,
     blockAccessListHash,
     slotNumber,
@@ -146,7 +141,6 @@ export function createBlockHeaderFromRPC(blockParams: JSONRPCBlock, options?: Bl
       nonce,
       baseFeePerGas,
       withdrawalsRoot,
-      parentBeaconBlockRoot,
       requestsHash,
       blockAccessListHash,
       slotNumber,